% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Entropy_All_4.R
\name{entropy}
\alias{entropy}
\title{Entropy Tests of Ordinal Patterns}
\usage{
entropy(p, ...)
}
\arguments{
\item{p}{a sequence of ordinal pattern probabilities}

\item{method}{string; The type of entropy or information to calculate}
}
\description{
Calculates Fisher Information or Shannon Entropy from Ordinal Pattern probabilities
}
\examples{
z <- rnorm(1000)
op.wn.z <- OPprob(z, emb=4)
entropy(op.wn.z, method="Shannon")
}
\keyword{Ordinal}
\keyword{Patterns}
