% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy_test.R
\docType{data}
\name{x}
\alias{x}
\title{Unit tests for Entropy function}
\format{
An object of class \code{numeric} of length 1000.
}
\usage{
x
}
\description{
File for running tests on the Entropy function
}
\keyword{Distributions}
