% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Entropy_All_4.R
\name{entropy}
\alias{entropy}
\title{Entropy Tests of Ordinal Patterns}
\usage{
entropy(p, ...)
}
\arguments{
\item{p}{a sequence of ordinal pattern probabilities}

\item{method}{string; The type of entropy or information to calculate; default "Fisher"}

\item{beta}{the parameter beta for the Rényi or Tsallis entropy; default 1.5}
}
\description{
Calculates Fisher Information or Shannon, Rényi or Tsallis Entropy from Ordinal Pattern probabilities
}
\examples{
z <- rnorm(1000)
op.wn.z <- OPprob2(z, emb=4)
entropy(op.wn.z, method="Fisher")
entropy(op.wn.z, method="Shannon")
entropy(op.wn.z, method="Renyi", beta=2)
entropy(op.wn.z, method="Tsallis")
}
\keyword{Ordinal}
\keyword{Patterns}
