% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPprob2.R
\name{OPprob2}
\alias{OPprob2}
\title{Function that finds the Ordinal Patterns probabilities in a time series for a given embedding dimension}
\usage{
OPprob2(TS, emb)
}
\arguments{
\item{TS}{time series of length \eqn{n-m+1}}

\item{emb}{embedding dimension \eqn{m}}
}
\value{
a sequence of \eqn{n} patterns
}
\description{
Function that finds the Ordinal Patterns probabilities in a time series for a given embedding dimension
}
\examples{
set.seed(1234567890, kind="Mersenne-Twister")
x <- rnorm(1000) # white noise
y <- mov.av(x, order=11) # smoothed with moving averages
OPprob2(x, emb=4)
OPprob2(y, emb=4)
}
